streamlit-app/requirements.txt
streamlit==1.32.2
openai==2.6.1
python-dotenv==1.0.1
streamlit-chat==0.1.1
streamlit-extras==0.4.0